more rust compatibility & convenience juggling
