+++ crates/store/re_types/src/lib.rs
@@ -280,5 +280,8 @@ pub mod image;
 pub mod tensor_data;
 pub mod view_coordinates;
 
+mod rotation3d;
+pub use rotation3d::Rotation3D;
+
 #[cfg(feature = "testing")]
 pub mod testing;
